refactor: std/pairing bls12377 api more coherent
